Milestone 1 Code
we have successfully completed several key tasks. Initially, we gathered the dataset, which includes emotionally labeled speech and audio files. We conducted an in-depth exploration to understand its structure and characteristics. Following this, we used the Librosa library to extract essential audio features, such as mel-frequency cepstral coefficients (MFCCs) and chroma features, which are crucial for emotion recognition. We then split the data into training, validation, and test sets to ensure the evaluation process.
With the dataset prepared, we moved on to the modeling phase. We selected the appropriate machine learning model (Librosa) that are well-suited for emotion recognition tasks. The initial model was implemented using the data, and we trained these models on the training dataset. Post-training, the models were tested on the validation set to assess their performance. This evaluation allowed us to identify areas for improvement and refine our models further.


These are the achievements and milestones reached
•	Gather the dataset.
•	Explore and understand the dataset structure.
•	Extract audio features using Librosa.
•	Split the data into training, validation, and test sets.
•	Select appropriate machine learning model for emotion recognition.
•	Implement initial model using data.
•	Train model on the training data
•	Test model on the validation set.
•	Evaluate initial model performance and identify areas for improvement.