Milestone 2 code
In this second installment we have successfully achieved several key tasks, in our progress we initially did cross validation to confirm that our model generalized well and was not overfitting. Likewise, we applied some tuning to the model using grid search, managing to improve the accuracy of the model. We also developed three new codes that we will use in the following stages, these are: 1. Downloading youtube audios, 2. Converting audios to mp3 files and 3. Splitting audios into small parts.

•	Fine-tune model hyperparameters.
•	Implement additional feature extraction techniques if needed.
•	Retrain and test models with refined parameters.
•	Perform cross-validation to ensure model robustness.
•	Compare performance metrics and select the best model.