Milestone 3 code
In this step on the first place we checked if the dataset was balance to try to improve our model performance, we realized the data was very lightly unbalance but still we decided to applied SMOTE to balance class distributions and used GridSearchCV for hyperparameter tuning of an MLP classifier. However, the model performed better without SMOTE, so we finally decided to leave the same model we had before, just improving the hyperparameters alpha to 0,001 and hidden layers to 500 we could improve our accuracy to 79,96%. 
Also We developed a function to generate and visualize spectrograms for different emotions and we analyzed them to gain insights. 


Milestone 3:
•	Check data balance
•	Finish tuning model
•	Create plots of spectrograms for different emotions.
•	Visualize spectrograms to analyze frequency patterns related to different emotions.